Embedding Lua Progress
Changes:

 - Reorganized SDF API to make this easier to implement.

 - Moved the C API bindings to c_sdf.cpp.

 - Lua environments are always recreated per model to prevent state from
   persisting.  This also helps catch memory management problems.

 - Lua API now implements __index on the SDF userdata type.

 - Hack: A copy of the model evaluator is passed into CompileEvaluator instead
   of the original to prevent a double free when the next model is loaded.

The above hack is a temporary measure to get other things working, and will not
be suitable once operators are exposed, as nodes in the tree are likely to be
GC'd in an arbitrary order.

The correct solution is probably some form of reference counting, however I
think it will suffice to just have some way to annotate a given SDFNode to
indicate whether it is safe for the high level language to delete it
automatically.

For example, if you construct a graph like so:

s = sphere(2)
model = union(
    s:move_x(-.5)
    s:move_x(.5))

Only the SDFNode wrapped by variable `s` should actually be deleted when GC
runs.  The SDFNode wrapped by variable `model` should only be deleted by
`ClearTreeEvaluator`, and its two children should only be deleted when their
parent's destructor is called.

So I think it should suffice to just add an ownership flag to SDFNode.