Refcounting, Assertions, and More Lua API Progress
SDFNodes are now refcounted.  Anything that holds a pointer should call the
::Hold method to increment the refcount, and ::Release to decrement it.  This
will also automatically delete the object when the refcount reaches zero.

It is safe to call delete on SDFNodes when the refcount is already zero.  This
will only happen if no hold has been acquired, which is useful for temporary
values with only one obvious owner.

This change also adds assertions to help validate the correctness of the
refcounting scheme.

Some loose ends:

 1. The Lua API is still missing rotation and paint functions as well as ray
    queries.

 2. The paint node will probably need to be reworked to be part of the brush
    and use the graph to propogate values.  I'll need to think on it.

 3. The C API will need to be updated for the new refcounting convention, and
    any other changes.

 4. Whatever changes are needed to allow Lua and Racket to embed side by side.