Library Build Now Includes Renderer Machinery, Sorta
This change adds almost all of the missing "Majescule" functionality to the
"Miniscule" build, essentially erasing the primary difference between them.

This is part of an effort to remove the embedded Racket runtime from Tangerine,
in favor of allowing the renderer to be invoked as a library.  To complete this
conversion, a C API will need to be exposed for starting, stopping, and
configuring the renderer.  The menu bar should also be hidden by default in
this mode, with the theory being that the user would prefer to access this
functionality programmatically.  There should also be some way to push CSG tree
handles to the Renderer for compilation, rather than the current model loading
approach.