Script Execution Environments
The idea here is that ScriptEnvironments will eventually be used to handle
event dispatching and managing rendering primitives.  A ScriptEnvironment
will also be able to spawn sub-environments that have roughly the same
capabilities.  The sub-environments would be executed in parallel in a thread
pool, so only languages that support creating multiple isolated thread-safe
environments are eligible for sub-environments.

This means that if I keep Racket as an embedded language, it is unlikley to be
available as a sub-environment.  However, a racket main environment would be
fine to launch sub-environments written in other languages.