Terminology Cleanup and RequestDraw Bugfix
This change primarily renames variables to clarify terminology, and also
features some light refactoring of the Emit* interface to consolidate two
of the functions.

This change also changes the RequestDraw condition, so that pending shader
compile jobs result in draws.  This is meant to resolve a problem that
sometimes occurs on more complex models in the interpreter mode, where not all
of the interpreter stack sizes are compiled on the first frame.