Added an update callback to Lua environments.
This adds a function to Lua called `set_advance_event` that lets the script
specify a callback method to be called every frame, like so:

```lua
set_advance_event(function (dt, elapsed)
	print("tempus fugit", dt, elapsed)
end)
```

I'm mostly just feeling out ideas right now.  I intend to eventually move the
execution of script environment(s) into a thread pool, and use a message
passing system of some sort to communicate between actors and the renderer, so
this will likely change quite a bit over time.

I'm thinking the presence of callbacks like this after the first execution of
the script will be the distinction of if a file is interactive, or just a
simple model, and thus whether or not the `model` global is meaningful.

But as of yet there's no other way to render anything so this isn't very useful
yet.