Local<->World Transforms Per Model, Ref Counting
From here it should be easy to allow for the Lua environment to create and
update "instances".  For now I'm thinking an "instance" should just be lies to
get the end-to-end behavior working correctly.

So it would work like this:

The Lua script calls `node:instance()`, which creates a userdata wrapper for
the SDFModel that handles the appropriate refcounting.  The Lua namespace for
the wrapper could be `tangerine.model` or `tangerine.instance`, it doesn't
really matter.

From there, I could build a transform machine on the instance, and then later
make it so that multiple instances can hold references to the same model
instead of redundantly constructing them.

I might want to swap out `std::vector<SDFModel*> LiveModels;` for a different
container type for easier element removal, but it probably doesn't matter for
now.