Fixed and Enabled Occlusion Culling :O!
I was using the wrong sampler for the depth pyramid, and it was likely reading
from the wrong mip map as a result.

With this active, the results are stable while the camera is idle, and
performance is improved substantially in several models, with no known
regressions.

The "brick spiral" model now runs at about 30 ms interpreted on a 2080 Ti,
which is a substantial improvement over its previous baseline of 83 ms.
Curiously w/ occlusion culling, the complied variant is only marginally faster
than the interpreted version now.

The "wheel" was less triumphant, dropping from 25 ms to 20 ms (zoomed in a
bit) w/ occlusion culling.  Compiling the shaders for the wheel drop it down
further to 7 ms.  I suspect there is less of a gain here due to octree bounds
being less tight due to the use of blending operators, and that the average
octree complexity is higher overall due to the tire tread construction.

The "seaside town" model nows runs at ~15 ms interpreted, vs the ~27 ms before.
When compiled, it now runs at about ~8 ms.

Simple models like "basic thing" don't show a noticable change in perf.

I suspect that tighter octree bounds might be more efficient now, and
eliminating negative space false positives will also be impactful for further
improving perf.

Some important areas to follow up on:

 1. There's currently no secondary draw stage for culling false negatives, so
    moving the camera causes voxels to disappear for a frame sometimes.

 2. Some models are more sensitive to camera movement than others.  It would be
    good to have a way to freeze the culling behavior to inspect from other
    views.  This would probably require sending a different view and matrix for
    culling to use.  It may be that we also want to be able to use the prior
    frame's view matrix for the first culling pass anyway, since that is what
    the depth corresponds to.

 3. An octree overdraw debug view would be very useful now, too.

 4. As noted 0e22d13ca350aea4f434c0e29070692d6b903d5b, the VS stage is doing
    the culling work redundantly per vertex.  It would be more ideal to use a
    mesh shader here, or a compute shader + glDrawArraysIndirect.  We could use
    an indirection buffer per model, and submit each draw call w/ a different
    offset into that buffer, then a culling compute shader would just need to
    have all of the voxel bounds in one buffer and loop over them.  It's
    probably still not great to submit a few thousand draw calls twice, but if
    we do some cleverness with the projection for the first culling stage, and
    accept every voxel that moved into frame this frame, then maybe we only
    need to submit a secondary conditional draw for moving objects.