Switching to Instanced Voxel Drawing, Sorta
This change switches to using glMultiDrawArraysIndirect to draw each voxel per
program variant in one go.  Each voxel has its own DrawArraysIndirectCommand
to draw exactly one instance.  This will be followed with a compute culling
pass to replace the current per-vertex culling logic.

This could be reworked further to put all of the voxels for all programs into
one buffer, and all draw command packets into another, so that culling can be
accomplished in a single dispatch, but this is fine for now.  In such a rework,
there would still be the same number of glMultiDrawArraysIndirect calls.

To avoid requiring GL_ARB_shader_draw_parameters, this change uses the base
vertex index to calculate the DrawID.