Compute Occlusion Culling
This doesn't seem to be worth it.  This costs 10 ms for the brick spiral level
than the current excelsior branch.

It is hard to get a good reading on what is going wrong here, as the sheer
number of events slows down nsight substantially :/

The culling behavior seems to be correct though, based on the behavior of the
freeze culling option.

It might be that this will work better if all of the culling can be performed
in one dispatch.  It might also be a good idea to rework this to use wave
intrinsics.