Ensure the UI is always drawn while exporting meshes.
Prior to this change, the UI would only be drawn if force redraw was active,
if a stat window was open, or as a response to an SDL event like a mouse move.

This change allows for partial redrawing in the idle case.  This skips drawing
or clearing the depth pass, which causes the previously rendered result to be
used.  The rest of the frame renders as normal.  This should prevent excessive
rendering while exporting while still looking the same.