Uniform Scaling Operator
A uniform scale factor may now be applied to SDF node trees.  While not
strictly necessary to do so, this is always implemented as a matrix transform.
In the future, it may be beneficial to update the TransformMachine struct to
support simple scaling on the input point if no other transformations are, or
something like that.  I figure just using matricies is fine for now, because
applying the scale operator is often an after thought, and implies intended
repetition with variation, so in practice it probably doesn't matter anyway.