Added Truncated Cone Brush
"truncated_cone" is too much of a mouthful to type, as are other variations on
the name, so this primative is simply called a "coninder" as it is not quite a
cone and not quite a cylinder.  You can also call it a "cylicone" if that is
more your speed.