Refactored Mouse Events, Advance Event
This consolidates some common functinoality, but the important thing is that
the mouse event callbacks just receive a single table as an argument, and the
mouse event and event callbacks now handle errors correctly.

The rational for passing a table instead of individual args to the mouse event
callbacks is that it makes exposing more parameters to the event in the future
easier (eg, screen coordinates), and it allows for the Lua application to
optionally use the same callback for several event types that might otherwise
have semi-incompatible args.  For example, I'll probably update paint.lua in
such once a way once I've added mouse move/enter/exit events.