Fixed Ray Miss When Ray Aligns Perfectly With Voxel Edge
This problem would occur at what felt like random screen resolutions, where the
conditions align such that a given pixel's view ray would line up exactly on
the edge of a voxel, causing an odd split in the object that would also be
further emphasized by the outliner.

This change prevents the issue by expanding the voxel bounds by an arbitrary
small amount, to widen clipping slightly.  The impact on rendering performance
is expected to be negligible, since this does not produce more overdraw, and
the new margin should not be significant enough to meaningfully affect ray
marching iterations.