WIP ImFileDialog Integration
This change provides the following:

 * Changes needed to build ImFileDialog and stb.
 * Updated Visual Studio project files and filters needed to build the above on Windows.
 * Replaced the OpenModel native file dialog code w/ ImFileDialog.

This does not include the required changes needed to build on Linux, and the export
dialogs have not yet been updated.

Additionally, with the default path set, it doesn't seem to be possible to navigate
downward in the file tree as I would expect.  I'm also not sold on the general aesthetics
of this widget, though it is impressive how easy it was to integrate.

I'm iffy on if I'll merge this branch, more experimentation is needed.