Fixed file open default directory behavior.
ImFileDialog requires an absolute path for the default directory, and behaves
strangely if you provide a relative path instead.

This change determines the absolute path from the executable directory instead
of the current working directory, so that part should be cwd invariant.  For
this decision to matter, the code for opening shader resources should also be
update to be made cwd invariant as well, but that is a project for another
commit.