Updated Exporters to use ImFileDialog, Connected Magica Voxel Exporter
This change re-implements the remaining export dialog interface popups, and
removes the need for gtk, etc.  This should also mean that it is possible to
export models now on Linux, although the build script still needs to be
updated.

I never actually connected the magica voxel exporter to the GUI, so this change
also makes it possible to export .vox files again.

The remaining TODO items for this branch are:

 1. Updated the bundled license information.
 2. Update the Linux build script for the new dependencies.