Partial Voxel Culling Experiment
This change modifies the VS occlusion culling test to only check the vertices
for the current triangle, instead of checking all of the verticies for that
voxel.

This greatly improves apparent culling perf around the periphery, and results
in about a millisecond of savings in all models tested.

I'm not going to be submitting this to the primary branch yet, as this does
produce some false negatives, which could result in strobing in some places
as there is no secondary drawing stage yet.

An interesting thing I found while working on this was that if I modified
this change to only do one tap on the averaged screen space vertex coordinate,
runtime perf improves by (iirc) about 4 or 5 milliseconds!  Unfortunately this
comes with more substantial false negatives.  I'll need to revisit this idea
at whatever point I finally add the secondary culling stage.

(All of the time observations above are with the interpreter)

Moving the occlusion culling step to a mesh or compute shader will be more
complicated with the partial approach, since the draws no longer are
all-or-nothing.