Updated Lua SDF APIs to Accept and Return Vectors Where Applicable, Updated Demos
Now most library functions can take a vector instead of three scalars when it
makes sense to do so.  This mostly removes the need for packing and unpacking
lists, but it is also useful for non-trivial model construction.

This change updates the queries, paint, and seaside_town demos to use the new
convention.  This is not strictly necessary, as these changes are backwards
compatible.