"Sodapop" Rendering Experiment WIP
This change is just some really uninteresting scaffolding to carve out a space
for me to later build a new experimental fast renderer for Tangerine.  This
also turns off vsync for now, to make it easier to validate that this renderer
can hit the performance target for VR later on.

The goal is to build a renderer that prioritizes minmaxing latency and speed.
To accomplish this the renderer will be built around feeding a stream of simple
triangles that are decorated only with vertex colors and interpolation.  The
triangles will be generated from meshing SDF models by some means.  My angle
here is that vertex attributes are a wonderful means to perform low frequency
object space rendering.  Robust light transports of nontrivial materials are
possible with this system without compromising either goals of speed and
minimal frame latency, because it can be performed asynchronously.  The
frequency in which vertices are scheduled for retracing can also be dynamically
adjusted to ensure that the frame stays within the budgeted time.  I think
this idea will also probably work best if the vertices of a given triangle
generally update in different frames.

For this experiment, the (re)tracing stage will probably occur continuously in
a thread pool on the CPU to simplify implementation.