Add Guix package
In the process, refine the Linux build.

The files `installation.h` and `installation.cpp` define a new
struct `TangerineInstallation` to encapsulate paths relative to the
executable. It helps with adapting to a more Unix-like directory
structure, with `bin` and `share` directories.

In `CMakeLists.txt`, support `cmake --install`, organize the file into
more targets, and eliminate the need to manually move `tangerine` next
to the `shaders` directory.