Cross Platform Exe Path
This change integrates the library "whereami" to determine the path of the
running executable.  See https://github.com/gpakosz/whereami for more details.

This allows us to have consistent behavior across all build configurations.

I also added a fallback case for when it fails on Linux, to add an additional
hint if the proc filesystem is missing.  This may be useful to help people
attempting to run Tangerine from an improperly configured chroot environment
without downgrading to behavior that might be less secure such as using
argv 0 as a fallback.