Helper Function(s) for Octree Building
This change adds the function `SDFNode::GatherBoxels` for extracting a set of
bounding boxes with an associated best resolution.  The intention is that this
will be used to construct octrees for meshing.  The bounding regions will be
used to determine where coverage might be needed, and the voxel size is meant
to determine the appropriate leaf node size.

After the octree is constructed, the leaf node corners should be evaluated to
determine the signs, and then full interior and full exterior leaves should be
pruned.  It may also make sense to prune the evaluator while populating the
octree, but such an optimization can come later after supporting benchmarks are
collected.

This setup process is intended to facilitate implementing the meshing algorithm
described in "Unconstrained Isosurface Extraction on Arbitrary Octrees" (2007).
https://hhoppe.com/proj/unconstrainediso/