WIP Primal Octree Implementation
This change introduces PrimalOctree, which is a drop in replacement for
SDFOctree that uses a primal approach for populating the octree: a leaf node is
retained only if the signs evaluated at its corners are not all equal.

This could be used as a basis for implementing the meshing algorithm described
in the paper "Unconstrained Isosurface Extraction on Arbitrary Octrees" (2007),
but this implementation has problems.

The first problem is that boxels are generated wrong for planar brushes.  They
should be ignored unless an operator elsewhere in the tree can constrain them.

The second problem is that the appending algorithm doesn't guarantee that
siblings keep compatible corner definitions.  This will nede some rework in
how the tree is created, but this is just a sketch.

The third problem(ish) is the current ideal voxel sizes are too coarse for
rendering with, and so are probably also too coarse for meshing.  We'll
probably want the divisor to be determined by the brush shape.

With this implementation, the "flower.lua" model produces an incorrect set of
bounding voxels, and any model incorperating planes (such as "shapes.lua")
gets stuck in an infinite loop until the grow algorithm runs out of memory.

Otherwise this seems to produce reasonable results so far.  This is also very
good at eliminating interior voxels, so while I was not intending on using this
for anything other than meshing, it might also accelerate the current renderer
too.