WIP Integrate / Rewrite "Surface Nets" 3rd Party Library
This started as an attempt to integrate one of these 3rd party libraries for
SDF mesh extraction via naive surface nets:

https://github.com/r03ert0/surface-nets
https://github.com/Q-Minh/naive-surface-nets

The "surface-nets" library has the advantage of being a simple stand-alone C
reimplementation of the javascript reference implementation for this article:
https://0fps.net/2012/07/12/smooth-voxel-terrain-part-2/

The "naive-surface-nets" library is a different implementation in C++ with the
stated goal of emphasizing clarity in the implementation.  It also appears to
provide concurrency via C++17's <execution> library, so it is likely more
performant than the "surface-nets" library.

However, the "naive-surface-nets" library has an unnecessary dependency on
libigl, which would be problematic for license compatibility with Tangerine.
The readme claims that libigl is only used for the interactive visualization
application, something I would tear out anyway.  The library also depends on
the "Eigen" vector math library, which was not mentioned in the readme.

In the interest in not spending my Sunday doing license research and tearing
out excessive 3rd party dependencies I opted to try to integrate
"surface-nets", found that there were a number of technical issues that I
disliked about the C implementation, and proceeded to fork it.

I'm about half way through the process of untangling the indexing scheme to
remove Javascript-specific workarounds, de-terse the C, and adjust conventions
to fit better with Tangerine; but my attention is starting to wane and I'm
starting to feel like this is a dead end.

I can probably power through and complete this in another sitting or so, but
it might be a better idea to just start over and fork "naive-surface-nets"
to extract the inner implementation and remove the unwanted dependencies.