Update Exporter to Use Naive Surface Nets Meshing Library
This this new version works great!  I'm so happy!  I need to rig up
something so the progress counters and abort button both work again, but this
is working great, and I expect that I'll be merging this branch back into
excelsior soon!

I think this library will be useful for unblocking my work on the new renderer,
too.