Hack Atomic Progress Controls Into Naive Surafce Nets Library
This change restores the atomic progress reporting feature to the mesh exporter, and
also restores the "good enough" and "halt" functions.  The "good enough" function
might not be useful anymore, so a future change will remove it from the mesh export
mode as well as the "refinement" step (which is no longer used).  The "good enough"
function is still likely to be useful for the point cloud export though.