Export Point Clouds No Longer Default PLY Export Behavior
This was originally the default behavior on the assumption that the meshing
would be performed in an external tool, on the basis that the built in meshing
functionality produced poor quality results.

Now that we are using the naive surface nets library for mesh export, this is
no longer a reasonable default.