Added --llvmpipe Option For Linux Builds
When set, this option sets the environment variable LIBGL_ALWAYS_SOFTWARE=1
before initializing SDL2 and OpenGL.  With luck, doing so opts the runtime
into using the llvmpipe driver, or some other form of software rendering.

This allows Tangerine to be used on a much wider range of hardware, although
the current renderer performs quite poorly in this mode.