Linux OpenGL driver check when glxinfo is available.
If glxinfo is available, this will attempt to check if the selected driver
would support at least OpenGL 4.2.  If it does not, then Tangerine will print a
message explaining it, and then attempt to opt into llvmpipe by setting the
environment variable "LIBGL_ALWAYS_SOFTWARE" to 1.

It would be nice to revisit this later to also check for extensions.