Experimental Rework of Repaint Behavior
This isn't quite right, but I think one idea here worth pulling
forward to a future attempt is the dirty var on SDFModel.  That
way, simply scheduling an advance doesn't need to require a
repaint, as we only care if the model changed presentation in
some way.  Likewise, this allows for handled UI events to
conditionally trigger repaints without making it explicit.

The event handling in Tangerine.cpp is wrong, but I don't think
I'm going to make meaningful headway on this implementation
while using my ancient laptop, because the ray marching renderer
is just that slow on this machine.