Drawable deduplication experiment
This change keeps a cache of existing drawables, and holds onto a
representative evaluator for comparison.  When a new SDF model is
created, it will attempt to re-use a matching drawable for the model
if one exists.

This means that if you instance an SDF node several times, said
instances will be able to re-use shaders and other common data between
them.

This does allow for deep comparisons, which might not be a good idea.

Finally, the cache is a vector instead of a map, because defining
an efficient SDFNode::operator< is non-trivial.

If I did this again (and I probably will), I'd probably only have used
a pointer comparison on the assumption that instancing models with
identical node graph almost always means they were instanced with the
exact same evaluator.

Anyways, this should help the convergence time on some systems.