Starting to port over the Sodapop renderer implementation
This is just some misc changes and the shaders.  It does not do
anything useful yet.

I'm planning on using the surface nets library for meshing, but maybe
with a different thread pooling behavior than the exporter, since the
exporter's method tends to totally saturate the CPU and degrade
interactivity.

The other oddity is that the SDFModel (not the SodapopDrawable) will
have to track some per-instance arrays that match the size of the
meshing result.  These are the vertex color buffers used to do object
space tracing and what not.

There will also have to be a thread pool for the continuous tracing
stuff, which we'd presumably want to shut down when the renderer is
not active.

Lots to do still.