Implementing the Sodapop renderer
This gets as far as meshing in a worker thread, and rendering the
results.  There's no system in place for populating the vertex colors
yet, so this just passes through the vertex positions as vertex colors
for now.

The meshing system isn't great.  This is re-using the naive surface
nets library that is used by the exporter, but does so in its own
thread.  Some models converge quickly, others take forever.  This is
not a good property to have here.

An adaptive incremental meshing system would be greatly preferred, but
this is ok for now.

Next up is probably populating vertex colors w/ colors evaluated from
the SDF models.