Minor improvements to Sodapop
This adds better messages for showing when async job start and
complete, and also adds a crude hueristic to ensure that most things
produce meshes in an acceptable time frame.

In place of a reasonable adaptive and iterative meshing algorithm, I
think this system could be improved by being able to re-schedule
meshing jobs targeting higher resolutions.  Subsequent jobs would
create finer LODs.  It would be nice if a mapping could be established
such that veretx colors could be copied between LODs, once there's a
system for object space rendering.  Maybe an octree representation of
each LOD's vertex list could be used for that, so that each vertex has
its own octree cell.

Alternatively we just establish the starting vertex colors after a
given meshing job finishes by evaluating the SDF graph.

Side note unrelated to the above: I wonder if trimming down the SDF
graph per vertex is useful for anything, or for the bounding sphere of
a vertice's connected edges or triangles.  My guess is probably not,
because by the time we know where the vertices are, we're probably
just going to do one more eval to extract the material parameters.