Sodapop Meshing Jobs Populate Vertex Colors
When a meshing job completes, it now populates the vertex color array
for its drawable.  This uses the evaluator to sample instead of the
octree, since the octree tends to be less accurate.  Neither always do
the right thing though.

There's no lighting stuff yet, but this is actually surprisingly good
never the less.  My "sigmar's garden" clone is now quite playable on
an old Core 2 Duo running entirely in llvmpipe!  Granted, runs at
single digit frame rates, but that is still enough for the UI to be
responsive.