OpenGL Context Creation Can Now Fall Back to OpenGL ES 2
This doesn't do anything useful yet because Dear ImGui will crash
right away, and presumably the rest of the renderer has to be
reworked, but it does create an OpenGL ES 2 context now if it fails to
create an OpenGL 4.2 context!

Because this isn't useful at the moment, I'm leaving the automatic
llvmpipe functionality in place for the moment.