Force Dear ImGui Backend into "ES2" Mode, use Glad loader
This change mods Dear ImGui's polyglot OpenGL3/ES2/ES3 to build as-if
it were for OpenGL ES2, and substitutes the library include for Glad
so that Tangerine is in control of GL initialization for both.

This appears to work just fine when running in OpenGL 4.2 mode.
Tangerine still crashes in ES2 mode, but Dear ImGui reports successful
initialization now.