Fixed Spurious Release Triggered by std::vector Resize
And some other stuff along the way.  I removed the octree from the
meshing job out of paranoia, but I think it might be safe to
reintroduce it.

Unfortunately this means I can't de-duplicate equivalent SDFs unless
they have the same pointer, until I work out a safer memoization
scheme.

But on the plus side, the crash bug appears to be gone now.