Added Thread Pool Fence Mechanism
Scheduler::Enqueue now takes an optional argument ("IsFence") which
when set causes the thread pool to perform a top-of-pipe flush.  This
will hold the inbox lock until all threads in the pool are waiting on
it.  The task will then be ran on only one thread before releasing the
lock.

This will be useful later for actions that require the thread pool to
be temporarily suspended while they run, such as updating data
structures that would be shared by multiple threads.