Fixed Early Deletion of Lua Models
Lua models that are created from the evaluator stored in the global
variable "model" are now held as a shared pointer for the life of the
script environment.  My earlier translation of everything to
std::shared_ptr did not correctly preserve that TreeEvaluator held a
reference implicitly upon creation, so because the corresponding
models wasn't actually stored anywhere in the new version, it just got
cleared immediately.

It seems like the renderer is mostly working now, however I was able
to trigger a segfault while reloading a script with lots of model
instances in it, so it seems there are still other problems with the
new scheme lurking about.