Fixed Sodapop crash when ShapeCompiler is the current renderer.
This was due to `std::static_pointer_cast` being used in Sodapop::Attach, when
it should have used `std::dynamic_pointer_cast` to confirm that the correct
drawable is available.  As a result, `VoxelDrawableShared` was getting recast
as `SodapopDrawableShared`, which in turn meant `Sodapop::Hammer` would hit
the sodapop-specific processing path and bad things happened.

For good measure, `WorkerThread` no longer calls `Sodapop::Hammer` at all if
the wrong renderer is active.

This change also deletes the copy assignment operator on `Buffer`, which is
unrelated to this bug, but I noticed it was incorrect anyway so here it is.