Disgusting Workaround for ImFileDialog Texture Use-After-Free Bug
The materials provided with ImFileDialog recommend calling glDeleteTextures
immediately in the DeleteTexture callback.  If you actually do that though you
will delete the texture before ImFileDialog and/or Dear ImGui is actually done
drawing with it.  The result is the icons being corrupted when you change
folders select your open / save target.  Also if Tangerine is running with the
"--debug-gl" commandline argument, then cout will also be spammed with error
messages for the offending frames.  Not great!

I had a look at the internals briefly, and I think it is plausible that
ImFileDialog was just written with some assumptions about Dear ImGui that are
no longer true in the version we're using now.

I think fixing this correctly in the library would be more trouble than it is
worth, so instead I added a crude deferred delete mechanism.  Now when the
glDeleteTextures callback is hit, we just record the texture handle to be
processed a frame after a file dialog is closed.  Works well enough for now.