Fixed Sodapop Reload Crash on NVidia
It feels obvious in retrospect, but models are not safe to destroy from the
worker thread pool.  This change is an effort to ensure that SDFModel
shared pointers always tick down to zero on the main thread, so that we don't
accidentally issue GL commands from the worker pool.

The thread pool only persists weak pointers, but it previously could destruct
a SDFModel on accident within Sodapop::Hammer if the refcount dropped to 1
while the weak pointer was locked.

To prevent this, UnloadAllModels is now called before creating a new script
environment, and this function has been upgraded to include a detach mechanism.
Sodapop::Detach uses a fence to safely remove the model's weakref from the
thread pool, and then releases the (hopfelly) final shared pointer on the main
thread afterwards.

Implementing Sodapop::Detach required adding an "unstoppable" task type to the
scehduler.  When a purge happens while one of these tasks is in flight, it'll
block until all of the worker threads are waiting for the inbox to unlock.  The
purge will then run the task once it is safe, move it to the outbox, and
eventually call the Done method without ever calling the Abort method.

I don't think this is a great solution, but the results are sufficient to make
Sodapop usable on my Windows machine.  I expect this probably also fixes latent
bugs on Linux, too.

In an abundance of caution, this change also tries to reduce the degree in
which we persist unecessary SDFModel weakrefs on the main thread as well,
although that is unlikely to have been problematic.

Anyways, if this approach (ensuring that std::shared_ptr refcounts tick down to
zero on the main thread only) proves to be too unweildly, a better alternative
will be to implement some sort of deferred delete mechanism, and then
destructors then can move unsafe resources into said queue.