Scheduler Sleeps When Idle Instead of Yielding Unless Sodapop is Active
This change prevents the CPU load from going to 90-100% while doing nothing.

When Sodapop is the current renderer, worker threads will repeatedly call
Sodapop::Hammer when idle, which ironically causes most of the pool to be
blocked when only one model instance is active.

When ShapeCompiler is the current renderer, the thread will now just sleep for
4ms when there are no jobs pending.

In the future, it might make sense to have some kind of per-thread "sleepy"
mode that can be toggled for more nuanced reasons.