Fix Access Violation in Faulty GL Finalizers
Prior to this change the finalizers used a [=] lambda capture that was intended
to implicitly copy the referenced various buffer IDs.  Instead, [=] does an
implicit capture of the `this` pointer, which was long deleted by the time the
finalizer would run when deletion is deferred to the main thread.

This change corrects the issue by making the by-copy capture explicit.