Disconnect LuaModel Callbacks Immediately Upon Garbage Collection
This change fixes a dangling pointer crash.

Prior to this change, `~LuaModel` would disconnect mouse callbacks
from the Lua registry, on the assumption that the reset in `SDFModelGC` would
be the one to trigger the model's destructor.  Because it is possible for a
`LuaModel` destructor to be called after the `LuaEnvironment` that created it
had been destroyed, this behavior could trigger a crash in Lua by attempting
to modify the Lua registry without a valid environment.

The fix is to not do any Lua API calls in `~LuaModel`.  This functionality now
lives in `LuaModel::OnGarbageCollected`, which is called immediately by the
garbage collection callback.  `~LuaModel` isn't needed anymore, but I added an
assertion as a reminder.