Added Continuous Concurrent Tasks & Scheduler Stats
This change replaces the sodapop-specific "hammer" mechanism with continuous
tasks, which are tasks that can reschedule themselves.  This change also adds
stats for the task queues to the perf stats window.

This change also bumps the default scheduler queue size to 2*20 elements to
make deadlocking improbable for current workloads, which carries the baseline
cost of about 16 megabytes for the four extant queues.  In the future I'd like
to move to a design where at least the inbox, outbox, and delete queues are
guanteed to drain, but there's a lot to do right now so that will have to wait.